requirements.txt
Flask>=2.0
Flask-SocketIO
python-dotenv
mysql-connector-python
eventlet